fix: Replacing floating point values by integers for prices (#33)
* feat: WIP replacing float by ints for prices

* feat: Replaced other occurrences of prices and resources

+++ database/bsgo/migrations/6_create_computers.up.sql
@@ -17,7 +17,7 @@ CREATE TABLE computer (
 CREATE TABLE computer_price (
   computer INTEGER NOT NULL,
   resource INTEGER NOT NULL,
-  cost NUMERIC(8, 2) NOT NULL,
+  cost INTEGER NOT NULL,
   PRIMARY KEY (computer, resource),
   FOREIGN KEY (computer) REFERENCES computer(id),
   FOREIGN KEY (resource) REFERENCES resource(id)
